#12 Use a milestone and lock for the entire Jenkins job
So that no two builds of the job can run in parallel. See
https://wiki.jenkins-ci.org/display/JENKINS/Lockable+Resources+Plugin .
If more than one builds of a job are waiting to start, all but the
youngest one are aborted automatically. This behaviour can be understood
as aggregating small changes into a larger build, as the modifications
of the older builds will be included in the Git history of the youngest
build. See
https://wiki.jenkins-ci.org/display/JENKINS/Pipeline+Milestone+Step+Plugin
.